chore(test-pass):make all tests pass
Following full es6 implementation, make all tests pass

[(Finishes) #165459854]
