+++ .nyc_output/processinfo/47ce88c2-e12a-4843-a0f7-cd20511ef4ef.json
@@ -1,0 +1,1 @@
+{"uuid":"47ce88c2-e12a-4843-a0f7-cd20511ef4ef","parent":null,"pid":11396,"argv":["C:\\Program Files\\nodejs\\node.exe","C:\\Users\\user\\Desktop\\meals\\banka\\node_modules\\mocha\\bin\\mocha","dist/server/test","--exit"],"execArgv":[],"cwd":"C:\\Users\\user\\Desktop\\meals\\banka","time":1555634509156,"ppid":8292,"root":"60a45733-3160-4ed3-8469-8718278e307d","coverageFilename":"C:\\Users\\user\\Desktop\\meals\\banka\\.nyc_output\\47ce88c2-e12a-4843-a0f7-cd20511ef4ef.json","files":[]}
