+++ .nyc_output/processinfo/4fa498c9-3d2a-4ea6-8fc6-8bb49af296ff.json
@@ -1,0 +1,1 @@
+{"uuid":"4fa498c9-3d2a-4ea6-8fc6-8bb49af296ff","parent":"47ce88c2-e12a-4843-a0f7-cd20511ef4ef","pid":10048,"argv":["C:\\Program Files\\nodejs\\node.exe","C:\\Users\\user\\Desktop\\meals\\banka\\node_modules\\mocha\\bin\\_mocha","dist/server/test","--exit","--no-config","--no-package","--no-opts","--diff","--extension","js","--reporter","spec","--slow","75","--timeout","2000","--ui","bdd"],"execArgv":[],"cwd":"C:\\Users\\user\\Desktop\\meals\\banka","time":1555634510843,"ppid":11396,"root":"60a45733-3160-4ed3-8469-8718278e307d","coverageFilename":"C:\\Users\\user\\Desktop\\meals\\banka\\.nyc_output\\4fa498c9-3d2a-4ea6-8fc6-8bb49af296ff.json","files":["C:\\Users\\user\\Desktop\\meals\\banka\\dist\\server\\test\\test-accounts.spec.js","C:\\Users\\user\\Desktop\\meals\\banka\\dist\\index.js","C:\\Users\\user\\Desktop\\meals\\banka\\dist\\server\\routers\\users.js","C:\\Users\\user\\Desktop\\meals\\banka\\dist\\server\\controllers\\users.js","C:\\Users\\user\\Desktop\\meals\\banka\\dist\\server\\models\\users.js","C:\\Users\\user\\Desktop\\meals\\banka\\dist\\server\\db\\userRecord.js","C:\\Users\\user\\Desktop\\meals\\banka\\dist\\server\\routers\\accounts.js","C:\\Users\\user\\Desktop\\meals\\banka\\dist\\server\\controllers\\accounts.js","C:\\Users\\user\\Desktop\\meals\\banka\\dist\\server\\models\\accounts.js","C:\\Users\\user\\Desktop\\meals\\banka\\dist\\server\\db\\accountRecord.js","C:\\Users\\user\\Desktop\\meals\\banka\\dist\\server\\routers\\transactions.js","C:\\Users\\user\\Desktop\\meals\\banka\\dist\\server\\controllers\\transactions.js","C:\\Users\\user\\Desktop\\meals\\banka\\dist\\server\\models\\transactions.js","C:\\Users\\user\\Desktop\\meals\\banka\\dist\\server\\db\\transactionRecord.js","C:\\Users\\user\\Desktop\\meals\\banka\\dist\\server\\db\\cashierRecord.js","C:\\Users\\user\\Desktop\\meals\\banka\\dist\\server\\test\\test-transactions.spec.js","C:\\Users\\user\\Desktop\\meals\\banka\\dist\\server\\test\\test-users.spec.js"]}
