--- .nyc_output/processinfo/58d495d9-1b3a-4731-a312-b644bff6980a.json
@@ -1,1 +1,0 @@
-{"uuid":"58d495d9-1b3a-4731-a312-b644bff6980a","parent":"f5e0023b-7d73-4f1d-b087-b2dcf14e57b2","pid":11980,"argv":["C:\\Program Files\\nodejs\\node.exe","C:\\Users\\user\\Desktop\\meals\\banka\\node_modules\\mocha\\bin\\_mocha","server/test","--exit","--no-config","--no-package","--no-opts","--diff","--extension","js","--reporter","spec","--slow","75","--timeout","2000","--ui","bdd"],"execArgv":[],"cwd":"C:\\Users\\user\\Desktop\\meals\\banka","time":1555598022618,"ppid":9772,"root":"1141a440-9c7d-44c8-941b-5ce6e292c012","coverageFilename":"C:\\Users\\user\\Desktop\\meals\\banka\\.nyc_output\\58d495d9-1b3a-4731-a312-b644bff6980a.json","files":["C:\\Users\\user\\Desktop\\meals\\banka\\server\\test\\test-accounts.spec.js","C:\\Users\\user\\Desktop\\meals\\banka\\index.js","C:\\Users\\user\\Desktop\\meals\\banka\\server\\routers\\users.js","C:\\Users\\user\\Desktop\\meals\\banka\\server\\controllers\\users.js","C:\\Users\\user\\Desktop\\meals\\banka\\server\\models\\users.js","C:\\Users\\user\\Desktop\\meals\\banka\\server\\routers\\accounts.js","C:\\Users\\user\\Desktop\\meals\\banka\\server\\controllers\\accounts.js","C:\\Users\\user\\Desktop\\meals\\banka\\server\\models\\accounts.js","C:\\Users\\user\\Desktop\\meals\\banka\\server\\routers\\transactions.js","C:\\Users\\user\\Desktop\\meals\\banka\\server\\controllers\\transactions.js","C:\\Users\\user\\Desktop\\meals\\banka\\server\\models\\transactions.js","C:\\Users\\user\\Desktop\\meals\\banka\\server\\test\\test-transactions.spec.js","C:\\Users\\user\\Desktop\\meals\\banka\\server\\test\\test-users.spec.js"]}
