--- .nyc_output/processinfo/f5e0023b-7d73-4f1d-b087-b2dcf14e57b2.json
@@ -1,1 +1,0 @@
-{"uuid":"f5e0023b-7d73-4f1d-b087-b2dcf14e57b2","parent":null,"pid":9772,"argv":["C:\\Program Files\\nodejs\\node.exe","C:\\Users\\user\\Desktop\\meals\\banka\\node_modules\\mocha\\bin\\mocha","server/test","--exit"],"execArgv":[],"cwd":"C:\\Users\\user\\Desktop\\meals\\banka","time":1555598020987,"ppid":1736,"root":"1141a440-9c7d-44c8-941b-5ce6e292c012","coverageFilename":"C:\\Users\\user\\Desktop\\meals\\banka\\.nyc_output\\f5e0023b-7d73-4f1d-b087-b2dcf14e57b2.json","files":[]}
