+++ .nyc_output/processinfo/index.json
@@ -1,1 +1,1 @@
-{"processes":{"58d495d9-1b3a-4731-a312-b644bff6980a":{"parent":"f5e0023b-7d73-4f1d-b087-b2dcf14e57b2","children":[]},"f5e0023b-7d73-4f1d-b087-b2dcf14e57b2":{"parent":null,"children":["58d495d9-1b3a-4731-a312-b644bff6980a"]}},"files":{"C:\\Users\\user\\Desktop\\meals\\banka\\server\\test\\test-accounts.spec.js":["58d495d9-1b3a-4731-a312-b644bff6980a"],"C:\\Users\\user\\Desktop\\meals\\banka\\index.js":["58d495d9-1b3a-4731-a312-b644bff6980a"],"C:\\Users\\user\\Desktop\\meals\\banka\\server\\routers\\users.js":["58d495d9-1b3a-4731-a312-b644bff6980a"],"C:\\Users\\user\\Desktop\\meals\\banka\\server\\controllers\\users.js":["58d495d9-1b3a-4731-a312-b644bff6980a"],"C:\\Users\\user\\Desktop\\meals\\banka\\server\\models\\users.js":["58d495d9-1b3a-4731-a312-b644bff6980a"],"C:\\Users\\user\\Desktop\\meals\\banka\\server\\routers\\accounts.js":["58d495d9-1b3a-4731-a312-b644bff6980a"],"C:\\Users\\user\\Desktop\\meals\\banka\\server\\controllers\\accounts.js":["58d495d9-1b3a-4731-a312-b644bff6980a"],"C:\\Users\\user\\Desktop\\meals\\banka\\server\\models\\accounts.js":["58d495d9-1b3a-4731-a312-b644bff6980a"],"C:\\Users\\user\\Desktop\\meals\\banka\\server\\routers\\transactions.js":["58d495d9-1b3a-4731-a312-b644bff6980a"],"C:\\Users\\user\\Desktop\\meals\\banka\\server\\controllers\\transactions.js":["58d495d9-1b3a-4731-a312-b644bff6980a"],"C:\\Users\\user\\Desktop\\meals\\banka\\server\\models\\transactions.js":["58d495d9-1b3a-4731-a312-b644bff6980a"],"C:\\Users\\user\\Desktop\\meals\\banka\\server\\test\\test-transactions.spec.js":["58d495d9-1b3a-4731-a312-b644bff6980a"],"C:\\Users\\user\\Desktop\\meals\\banka\\server\\test\\test-users.spec.js":["58d495d9-1b3a-4731-a312-b644bff6980a"]},"externalIds":{}}
+{"processes":{"47ce88c2-e12a-4843-a0f7-cd20511ef4ef":{"parent":null,"children":["4fa498c9-3d2a-4ea6-8fc6-8bb49af296ff"]},"4fa498c9-3d2a-4ea6-8fc6-8bb49af296ff":{"parent":"47ce88c2-e12a-4843-a0f7-cd20511ef4ef","children":[]}},"files":{"C:\\Users\\user\\Desktop\\meals\\banka\\dist\\server\\test\\test-accounts.spec.js":["4fa498c9-3d2a-4ea6-8fc6-8bb49af296ff"],"C:\\Users\\user\\Desktop\\meals\\banka\\dist\\index.js":["4fa498c9-3d2a-4ea6-8fc6-8bb49af296ff"],"C:\\Users\\user\\Desktop\\meals\\banka\\dist\\server\\routers\\users.js":["4fa498c9-3d2a-4ea6-8fc6-8bb49af296ff"],"C:\\Users\\user\\Desktop\\meals\\banka\\dist\\server\\controllers\\users.js":["4fa498c9-3d2a-4ea6-8fc6-8bb49af296ff"],"C:\\Users\\user\\Desktop\\meals\\banka\\dist\\server\\models\\users.js":["4fa498c9-3d2a-4ea6-8fc6-8bb49af296ff"],"C:\\Users\\user\\Desktop\\meals\\banka\\dist\\server\\db\\userRecord.js":["4fa498c9-3d2a-4ea6-8fc6-8bb49af296ff"],"C:\\Users\\user\\Desktop\\meals\\banka\\dist\\server\\routers\\accounts.js":["4fa498c9-3d2a-4ea6-8fc6-8bb49af296ff"],"C:\\Users\\user\\Desktop\\meals\\banka\\dist\\server\\controllers\\accounts.js":["4fa498c9-3d2a-4ea6-8fc6-8bb49af296ff"],"C:\\Users\\user\\Desktop\\meals\\banka\\dist\\server\\models\\accounts.js":["4fa498c9-3d2a-4ea6-8fc6-8bb49af296ff"],"C:\\Users\\user\\Desktop\\meals\\banka\\dist\\server\\db\\accountRecord.js":["4fa498c9-3d2a-4ea6-8fc6-8bb49af296ff"],"C:\\Users\\user\\Desktop\\meals\\banka\\dist\\server\\routers\\transactions.js":["4fa498c9-3d2a-4ea6-8fc6-8bb49af296ff"],"C:\\Users\\user\\Desktop\\meals\\banka\\dist\\server\\controllers\\transactions.js":["4fa498c9-3d2a-4ea6-8fc6-8bb49af296ff"],"C:\\Users\\user\\Desktop\\meals\\banka\\dist\\server\\models\\transactions.js":["4fa498c9-3d2a-4ea6-8fc6-8bb49af296ff"],"C:\\Users\\user\\Desktop\\meals\\banka\\dist\\server\\db\\transactionRecord.js":["4fa498c9-3d2a-4ea6-8fc6-8bb49af296ff"],"C:\\Users\\user\\Desktop\\meals\\banka\\dist\\server\\db\\cashierRecord.js":["4fa498c9-3d2a-4ea6-8fc6-8bb49af296ff"],"C:\\Users\\user\\Desktop\\meals\\banka\\dist\\server\\test\\test-transactions.spec.js":["4fa498c9-3d2a-4ea6-8fc6-8bb49af296ff"],"C:\\Users\\user\\Desktop\\meals\\banka\\dist\\server\\test\\test-users.spec.js":["4fa498c9-3d2a-4ea6-8fc6-8bb49af296ff"]},"externalIds":{}}
