+++ src/server/db/cashierRecord.js
@@ -1,0 +1,20 @@
+const cashierRecord = [
+  {
+    id: 1,
+    firstName: 'AnalyserNode',
+    lastName: 'TreeWalker',
+  },
+  {
+    id: 2,
+    firstName: 'Analyser',
+    lastName: 'Tree',
+  },
+  {
+    id: 3,
+    firstName: 'Node',
+    lastName: 'Walker',
+  },
+];
+
+
+export default cashierRecord;
