+++ src/server/db/userRecord.js
@@ -1,0 +1,15 @@
+const userRecord = [{
+  id: 1,
+  email: 'dami@gmail.com',
+  firstName: 'Dami',
+  lastName: 'Akande',
+  password: '123456',
+  type: 'staff',
+  isAdmin: false,
+  sex: 'M',
+  mobile: '1234567890',
+  active: false,
+}];
+
+
+export default userRecord;
