+++ src/server/routers/transactions.js
@@ -1,5 +1,5 @@
 import express from 'express';
-import { TransactionController } from '../controllers/transactions';
+import TransactionController from '../controllers/transactions';
 
 const router = express.Router();
