+++ src/server/test/test-accounts.spec.js
@@ -3,7 +3,7 @@
 import chai from 'chai';
 import chaiHttp from 'chai-http';
 import server from '../../index';
-import { AccountController } from '../controllers/accounts';
+import AccountController from '../controllers/accounts';
 
 // eslint-disable-next-line no-unused-vars
 const should = chai.should();
